Updated openstack/openstack Project: openstack-infra/puppet-httpd  f0d478b4f9a64ade24ed3b1ee6123607268ea52e
Add logrotate class to Apache

Adding this subclass to httpd configuration, to allow
modules using this manifest to setup log rotation for
apache on demand.
Please note that apache is installing a logrotate file by
default here, but naming that on same way will overwrite it,
and allow our desired log rotation.

Depends-On: I7451ecc550b50b85c55784ef598f6b45412117fc
Change-Id: I3ef60fd9386795007e5b55264cc6e91fa0b67c68