Updated openstack/openstack Project: openstack/neutron  5384047f72fc6968f53e3abd4e55837423fde11a
Only get host data for floating ips on DVR routers

First, we are only interested in setting the host attribute when
the router is a DVR router.  Second, we don't need to query all of
the ports on the host, we just need to query the ports that are
referenced by the set of floating ips that we already have.

Closes-bug: #1496974
Change-Id: If611de14b2ab77d2eb9ce8c5b307ea6dd4403fe1