Updated openstack/openstack Project: openstack-infra/subunit2sql  a2c496f77524db674b56c5d42b612faf548f0a2e
Add tests for get_tests_run_dicts_from_run_id() and fix issues

This commit adds unit tests for the get_tests_run_dicts_from_run_id()
db api method. It adds tests for a variety of different data scenarios
where different pieces of data could be missing.

In the process of adding the tests a number of bugs with the function
were uncovered. If there was no test run metadata for a particular
test run a result wouldn't be returned for that test run because of
the use of an inner join. If either timestamp field was missing the
method would fail because it would attempt to set the microseconds
on a NoneType object. All of these were addressed as part of this
commit to enable tests to work.

Change-Id: Ifde4158137fd4e985ce9f38608cea553348cd2aa