Updated openstack/openstack Project: openstack/heat-cfntools  6571e5ab646f55ba9411f5a668c5224203a08707
Don't run commands given as list on shell

Commands from AWS::CloudFormation::Init, when supplied as list, should
be run with shell=False. Only when commands are given as string, they
are meant to be run on shell.

In principle, we are trying to give least access to the shell to avoid
any inadvertent shell injections.

Change-Id: I3dc6fe0c29a14f75be044846f737e1ade23a6d6b
Closes-Bug: 1498300