Updated openstack/openstack Project: openstack/cinder  8de60a8b40a32921d2d93fea828e73dcfcd2931b
Fixing create CG from Cgsnapshot bug in VNX driver

When creating a consistency group from a cgsnapshot, VNX driver
didn't force a poll to the VNX system. This caused the creation
to fail because the out-of-date status was returned. This patch
forces a poll to the VNX system to get rid of this issue.

Change-Id: I710bd0ac2a6049a79fce0d552ad8fba2c9fbde67
Closes-Bug: #1496660