Updated openstack/openstack Project: openstack/trove  39253df463573b7f22314d4ea54adbcdddaaae7a
Fix Mongo report_root call to have correct args

report_root is being called in the case of a new instance being
created from backup when root is enabled on the originating instance.
A change [1] was made that added the context object to the report_root
argument list and the Mongo guest wasn't passing it.

[1] https://review.openstack.org/#/c/124355/

Change-Id: Id61337ae85fb980752d606bddb2f1e574a345cba
Closes-bug: 1498568