Updated openstack/openstack Project: openstack-infra/project-config  200cbfed19efdfe905b42cbecb2c3156b931d514
Retire the overcloud-f21-nonha job

Now that it is identical to overcloud-f21puppet-nonha
it can be removed. All of the tripleo jobs are now puppet
based so their is no need for a seperate check-tripleo-puppet-jobs
template.

The result of this is that all of the tripleo projects run
a combination of
  - gate-tripleo-ironic-overcloud-f21puppet-ceph
  - gate-tripleo-ironic-overcloud-f21puppet-ha
  - gate-tripleo-ironic-overcloud-f21puppet-nonha

While others that do not require the rull suite of tests get only
the nonha job.

Change-Id: Iac6e6c62135d1dc8d91dbdc82b27dd7d3267e1b3