Updated openstack/openstack Project: openstack-infra/system-config  35694901b0b21d9791eb414626ddf0fae560e1d7
Staged upgrade for npm

The version of npm that ships with trusty is able to upgrade itself
to npm v2, but cannot upgrade to npm v3. Furthermore, performing
a staged upgrade picks up the default `profile` parameter from
npm, so that value has to be overridden globally in order to ensure
that /usr/bin is not polluted.

This patch adds the following:
- A staged upgrade step, triggered when the detected version of npm
  is "1", to upgrade to version 2.
- A new global npmrc configuration file, containing the profile setting.
- Renamed jenkins/global npmrc templates.

Values that made more sense to keep globally have been moved there.

Change-Id: Ia633ccabb49d74a10c904a6e8ba73945d77fb32d