Updated openstack/openstack Project: openstack/api-site  f4b2b4871ed7d3a49497f0b8080c88a7a67e5dec
Add two missed optional request parameters

There are two optional request parameters missed in
"Create a load balancer health monitor" call as below:
"http_method (Optional)"and "admin_state_up(Optional)".

there also is a mistake for the request parameter "tenant_id".
It should be "Optional",but now it's not.
and the description of "tenant_id" has mistake for VIP.

I added the two request parameters and correct the mistakes.
Closes-Bug:1495229

Change-Id: Ib6263d4770007e0e729ab085e64a2ba23d06a4a9