Updated openstack/openstack Project: openstack/ceilometer  c362758958cbbc95977f08f3ca7c03c2a3a8cf71
Coordinator handles ToozError when joining group

While joining partitioning group, we check for
MemberAlreadyExist and GroupNotCreated, but fail
to account for connection failure scenarios. This
leads to ugly stack trace in notification agent logs
and bails out before configuring the notification
and pipeline listeners.

This fix handles the ToozError exception and logs
an error message before retrying the join, assuming
that the error was a temporary problem.

Change-Id: I2aed2241ded798464089b3eec5e1394422a45844
Closes-Bug: 1496982