Updated openstack/openstack Project: openstack/nova  442501534eecea61b448f5e314dbed7d40947e5c
Cells: Limit instances pulled in _heal_instances

In order to avoid ballooning memory usage we should limit the number of
instances pulled at any time while syncing them.  This adds pagination
to the generator used in order to limit it, and an optional parameter to
optimize for when small numbers of instances are needed.

Change-Id: Ia82185af27ed11d68a8498ec4bde3da529134893
Closes-Bug: 1447732