Updated openstack/openstack Project: openstack/tripleo-puppet-elements  7d7599b48085dac10c8bfbf2c47043d74511f4d8
Add puppet-healthcheck to the list of modules

The puppet-healthcheck module comes with a provider tcp_conn_validator
that allows one to make sure a port is up and running before the catalog
application move onward. This will be usefull in tripleo-heat-template
to avoid sleep with a random number of seconds.

Change-Id: I1abf7a52a0696e392fdb85a8adca7ddcf81727b6