Updated openstack/openstack Project: openstack/api-site  2c71d4540291bc3b9235bb598e41cde155fbff8f
Correct description of response parameters
Some description of response parameters  are incorrect
for calls of "health monitors"(getlist,show,create,update).
"The unique ID for the VIP." should be as below:
"The unique ID for the Health Monitor"
"The ID of the tenant who owns the VIP..." should be as below:
"The ID of the tenant who owns the Health Monitor..."

So I correct this in /api-ref/src/wadls/netconn-api/src/common.ent.
Closes-Bug: #1493702

Change-Id: Iab34fefe3e65eb2c56bd686724feb6f08857cd51