Updated openstack/openstack Project: openstack/openstack-manuals  c074d04becf363c3b8792341c36319badac7e87e
Add Ironic Section in Administrator Guide

Since Ironic is an integration project now. We should add one section
about it in Admin guide.
This commit is part of the bugfix since adding the whole section is
huge. This one is to create a new file for ironic section and add
basic outline.

Change-Id: I355245629ec46be4a44aef6dea33c60306edbb6d
Partial-Bug: #1495898