Updated openstack/openstack Project: openstack/heat  8f1ff524f88d14672a1b99262cd9c0428ffbc07b
Only fail Volume deletion in 'error_deleting' state

The volume may have been in the 'error' status before we attempted to
delete it, and there is no guarantee that it has moved to 'deleting' before
we query it again. We only want to fail if there is an error *after* we
call delete, which will put the volume into the 'error_deleting' state. We
now check for that state explicitly.

Change-Id: I7e205aef46f8dd35fd605eb768a3b05360f4d374
Closes-Bug: #1496445