Updated openstack/openstack Project: openstack/stevedore  e327c8226fe2589c0cf5f688941f16aac87edde2
Show how to add a plugin in a separate package

One of the most common misconceptions about plugins is that they must be
installed to the same python package as the application. Update the
documentation to provide an explicit example of providing plugins from a
separate package.

Also clean up the package metadata for the existing example to remove
references to URLs and email addresses that no longer work.

Change-Id: I801b6ec3ff287c5342ce63a865167f55d31c78d3
Closes-Bug: #1394393