Updated openstack/openstack Project: openstack-infra/project-config  4293826b476d8dbbfc1e5ad9fe86e6a8d0a1b747
Add a new script that queries for a Zanata version

To allow us to check if a version exists for a Zanata project without
requiring us to create a zanata.xml first, add a small script that
uses the Zanata REST API to query for its existance.

Make use it instead of the zanata-cli in upstream_translation_update,
since we can't create a zanata.xml if the version doesn't exist.

Change-Id: I93a72845aa705ce6b72acec9b5d62b5565368ab0