Updated openstack/openstack Project: openstack/zaqar  97bbbfbb3506613fd76ab4f52adb7d07469dd7aa
Cleanup of Translations

In preparation for the release, do some cleanups for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

This updates also recreates all pot (translation source files) to
reflect the state of the repository.

This change needs to be done manually since the automatic import does
not handle some of these cases.

Change-Id: I7024b3bf07c5db670e8a680ef9f27b5fea45d1a0