Updated openstack/openstack Project: openstack/nova  fd64ddb46c4954730374761169b97c3046eb8d3d
Revert "Deprecate cinder.cross_az_attach option"

This reverts commit d8d2f02f3014e8d7e0ef9a61723f5a10f050cc81

The deprecation change was merged before operators had a
chance to weigh in and it turns out that at least one large
operator relies on this option for isolating VMs and volumes
to the same availability zones.

There are still a lot of issues around the usage of this
option and the tightly coupled interaction between nova
and cinder with respect to availability zones.  We need
to better document this option and how it's used which
will come in a follow up change.

We also need to address the usability issues around this
when doing a boot from volume where source != volume,
because right now the failure happens when building the
VM and that results in a NoValidHost error for the user
if they didn't specify the request correctly or their
admin doesn't have the cloud configured correctly.

Ultimately there is a lot of API policy defined between
this option and the allow_availability_zone_fallback
option in cinder and that's hidden from the user. So there
will probably also be a follow up change to add validation
in the compute API in the boot from volume case where
source != volume and cinder.cross_az_attach=False so
we can attempt to fail fast with a 400 response rather
than get to the compute node and fail the build.

There is related discussion in the mailing list here:

https://www.mail-archive.com/openstack-dev@lists.openstack.org/msg64179.html

Change-Id: Ia6c8200286242d637f409ce0e65d906b5ce4b952