Updated openstack/openstack Project: openstack/horizon  52cca1d21bfa8c6f06c027fa513d388d5f1a546e
Cleanup of Translations

In preparation for the release, do some cleanups for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

This change needs to be done manually since the automatic import does
not handle some of these cases.

Change-Id: I8717e6a067f529e159cbefbc84fe308947b438cd