Updated openstack/openstack Project: openstack/neutron  b10e3bc46fa5f5e870c84b65b672e2fb69cc2b4b
SimpleInterfaceMonitor: get rid of self.data_received flag

It's not used anywhere outside tests, and there are better ways to
wait for updates.

The flag was once used to influence the state of activeness for the
monitor, but not anymore [1].

This cleanup also allows us to remove custom _read_stdout from the
monitor and reuse the one inherited from AsyncProcess. Meaning, we
also can safely get rid of another pile of duplicate tests.

[1]: I05faeddd061ab45af51c044a10462c3a57593d4d

Change-Id: I612d492c8f65a70e18f782a5e4d055de5f7948ef