Updated openstack/openstack Project: openstack-dev/devstack  9f6b7eb3e5faeed3bb54f8ba919bd9412fa8f7c7
attempt to cut api workers in half

One of the key reasons for the large number of API_WORKERS was that
mysql would block API workers, so would start rejecting work. Now with
the python mysql driver we should be eventlet aware, and life should
be good.

Let's see if this works.

Change-Id: Iaf8730a4dcdc30ef390795bfb5fb73ec3cd665fe