Updated openstack/openstack Project: openstack/python-barbicanclient  2e20dde5f322c480bcce7b43ef154c0cb8d3af51
enable barbican help without authentication

This fix enables 'barbican help' and 'barbican complete' to
not need authentication. This commit also enables deferred_help
so that the following command can be ran:
'barbican secret list --help' (the help for the
particular command will be shown).
The code changes are similar to the python-openstackclient
usage of the cliff library. The current implementation
of the barbican client authenticates before checking the type of
command. This early authentication causes the client
to fail with 'barbican help'. This code change will
check if the command requires authentication
and creates the client.

Change-Id: If335ef0373c75b145dc696d8501153da59fff1b7
Closes-Bug: #1488934