Updated openstack/openstack Project: openstack/python-openstackclient  678e69064854e3a3d3499171b0a29f30f2771840
Override the debug default and help text

Cliff sets the default debug value to False and this makes it
impossible to override debug with OCC.  If we set the default to
None, we can override debug in clouds.yaml.  Also, OSC changes
the meaning of --debug, so modify the help text.

Change-Id: I5e6680b2286cd7f55afe4b083fae5f8a4a9567a2
Closes-Bug: #1483378