Updated openstack/openstack Project: openstack/tempest-lib  fc6ea99b77e043d84d69ee35f8f4ffc363ac34c3
Switch to using autogenerated ChangeLog in docs

This commit switches the docs to use the PBR generated ChangeLog
instead of a hand curated release notes section. This is a lot less
error prone and instantaneous as compared to our current method which
often leaves the release notes section stale for months at a time.

Change-Id: Idbbe0d4ec8013ead19994056553f5ed1fcb49132