Updated openstack/openstack Project: openstack/python-openstackclient  83282bc5e133d8cd1718df524c1fa06add130b8a
attempt to find resource by listing

add a last-ditch effort to find the resource in question by
listing all the resources and doing a simply match for name and
id. if no match is found then raise an error, if the list call
is unsuccessful, raise the same error. we have failed this city.

Closes-Bug: #1501362

Change-Id: I0d3d7002e9ac47b17b1ef1a5534406c85b1fc753