Updated openstack/openstack Project: openstack/tempest-lib  ecbef55152144d8302024cadbb66eda716036e42
Add exception message to invalid cred field

This commit adds a new error message to the InvalidCredentials
exception raised when creating a new credentials object. Previously
it would just fail saying invalid credentials which wasn't
particularly useful for debugging what about the fields passed in was
invalid. The new error message will inform the user which key was the
cause of the error.

Change-Id: I49941321db7d6259080d706b687ac27173a3bffe