Updated openstack/openstack Project: openstack/stevedore  5c28a12b30198550e34be1832f25d23ce38bf87e
replace the hard-coded history list with an auto-generated one

Replace the hand curated list of changes with the one generated by pbr
automatically.

At the same time, change the toctree settings in index.rst to avoid
showing every release of stevedore at the top level and to clean up the
amount of nesting showing for other entries.

Change-Id: I3fb9306c0ca3ce6ed9681ffeaa8b9b8ed0f2bd7c