Updated openstack/openstack Project: openstack/horizon  258128006bb7bb17929b8805c0dd6c7718341f75
Neutron Quota Settings Flag Disables Neutron GUI

if OPENSTACK_NEUTRON_NETWORK["enable_quotas"] = False buttons will not appear
and exceptions are thrown.   This is because a dict is searched for a key
("available") which wont be there if its disabled.  This generates a
KeyError at runtime.

As currently written enable_quatas -> False essentially disables Neutron usage.
This is not a desirable outcome.

This patch shows Neutron related buttons and ignores the quota settings,
which seems more in line with the notion of disabling quotas.

Closes-Bug:#1482354
Change-Id: I9485ddeece74d87317c2587498f5f79d4ffdb66e