Updated openstack/openstack Project: openstack/ironic  e75b51d6ef0253d975c2d596944d0ec637a35cd5
Cached file should not be deleted if time equal to master

The cached file was being deleted when the modification times were
identical. This causes an issue especially with local storage of the
master images as a hard link is done and the cache file would always be
flagged as out of date.

Also give more details when the cached file is deleted.

Closes-Bug: #1499601
Change-Id: I0964c84028271b8a11e1178463bb2d513ce26410