Updated openstack/openstack Project: openstack/ironic  4f5651de18b1d13ec70f4fce6f7f0e651fb91af2
Make generation of ironic.conf.sample deterministic

This commit makes the generation of ironic.conf.sample
more deterministic by sorting the options belonging
to the same group based on module in which they are
defined before printing them.

It also fixes a small bug which caused options to
appears in different group in sample file generated
when they are the same (like [cisco_ucs]max_retry and
[cimc]max_retry which are of same type, have same default
value and same description).

Change-Id: Ifd336e2ec680a286e8789c2046bda2cb9ddbaf08
Closes-Bug: 1498113