Updated openstack/openstack Project: openstack-infra/subunit2sql  3563997bdbc5e0661d2a03bcd66992cc9e29e027
Switch to using autogenerated ChangeLog in docs

This commit switches the docs to use the PBR generated ChangeLog
instead of a hand curated release notes section. This is a lot less
error prone and instantaneous as compared to our current method which
often leaves the release notes section stale for months at a time.

Change-Id: If7354fc0612ed6d4dd76c8c595ba4722367fe65e