Updated openstack/openstack Project: openstack/python-glanceclient  5eaa52eae476ad1208b9156af558f6a4f4277119
No auth when token and endpoint are passed

The latest change to the auth_required logic introduced a bug were even
when the token and endpoint were passed, authentication was being
required.

This patch fixes that issue and makes sure that authentication is not
required when these 2 arguements are present.

Closes-bug: #1499540

Change-Id: I4c9c15ba526378970da5461511ed922d42c5a9f9