Updated openstack/openstack Project: openstack-infra/puppet-zanata  9453528beeb1b9b5ed8794f9fdc102d3dd045eb9
Bump java_xmx from 2G to 4G

The java_xmx value in Puppet for the -Xmx JAVA_OPTS variable was
increased 10 days ago following a timeout on the dev server with an
upload.  Ran into it again during a large branching of Horizon
translations, so now increasing it to 4G following analysis of
server usage to confirm it can take the load.

Change-Id: Iee63c91456ad8f17b0a32732f6610416b6c5b987