Updated openstack/openstack Project: openstack/oslo.config  d00c38c3673aa8cf92de6b5d818f84ab4659cefb
Correct the comment of __call__ method

The comment of return value is incorrect
in class ConfigOpt's _call_ method.
Now the comment of return value is described as below:
"returns: the list of arguments left over after parsing options"
but,actually, the method should not have return value.
So I delete the comment of return value.

Change-Id: I9bf5a412cf0a6f9ad70ea373716279154975b85d
Closes-Bug:1498378