Updated openstack/openstack Project: openstack/barbican  5c07034a7d9797313f2571b442d5d4c1312b6356
Cleanup of Translations

In preparation for the release, do some cleanups for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

This updates also recreates pot (translation source files) to
reflect the state of the repository in case there was no recent
import for them.

This change needs to be done manually since the automatic import does
not handle some of these cases.

Change-Id: I6b436ddc04f2f866dcb795d4d7c1c44d816dcadf