Updated openstack/openstack Project: openstack/barbican  4afaee095a78cbc473278ac42bb2f9de949af0a8
Check a CA's status as project and preferred CA before deleting

If a CA is preferred and not the last CA of a project, it should
not be deleted.  A user is informed of this with a 409 status code.

Otherwise, the CA can be deleted as well as any record of it in the
CA project list and the CA preferred list.

Change-Id: I9a1ee91252ee17746cfcffd11cba520270d09f21
Closes-bug: #1499876