Updated openstack/openstack Project: openstack/manila  2d39dba571a8d282802c5294682c467fe2f33699
Remove language about future features from driver doc

The community agreed it isn't a good idea to promise future driver
features in the share_back_ends_feature_support_mapping document.
This commit removes the implication that future features may be in
the document, and it removes one reference to Mitaka in one driver.

Change-Id: I953cb92edfca3bd54f3196d60b9e80589b5033d1