Updated openstack/openstack Project: openstack/barbican  f03dc6006edd17124c4ff5fbec5d3fb9ee6ffa01
Catch exceptions raised by bad certificate plugin info

If the Snakeoil plugin's persistent storage gets hosed or out of sync with the
Barbican database, the plugin can send information that raises an
exception.  This exception needs to be caught.  The failure is in
the plugin, so Barbican can not take guesses.  It should log the
error, cancel the action, and continue.

Change-Id: Id55eafa0dea2004bf2ab94aadb820bf08bec1468
Closes-bug: #1499768
Relates-to: #1498717