Updated openstack/openstack Project: openstack/oslo.middleware  20ef38112c1b565b698d54916a60550cc8ea491f
Allow the healthcheck middleware to provide more detailed responses

It can be quite useful to have admin endpoints for healthcheck
middleware provide more information about the health of the system.

To enable this kind of information this initial step starts to
add a jinja2 template that will produce the existing simple case
by default, but when setup with detailed mode it can then produce
more information about which checks were performed and there results.

It can be expanded in the future to provide more information as
desired (possibly showing more useful application level info when
in detailed mode).

Closes-Bug: #1494531

Change-Id: I9c6dae848985d950c3cf6798695c7669eb245ccb