Updated openstack/openstack Project: openstack/oslo.middleware  ecd79a5811cf40b59448bd210113357095f22842
Further improve the healthcheck output

Add on information about the active threads and active
greenthreads and some basic server information to be able
to show this to operators (or others) that have detailed
mode enabled and can benefit from this information being
output.

This addition makes it so that basic diagnostic information
can be obtained from the response of the healthcheck middleware
and not just OK or NOT OK; making the eventlet backdoor feature
that oslo.service provides less needed for operators.

Change-Id: I469c56586597561633921ebf3916585646f45373