Updated openstack/openstack Project: openstack/cinder  fb45dc953b6146810374499b3b5a193e478f6d29
Huawei driver add check before use a QoS

One QoS in Huawei storage array can contain at most 64 LUNs.
Do a check before adding LUN to QoS, if the number of the
LUNs in the QoS is already equal or larger than 64, then
this QoS should not be used.

Closed-Bug: #1498773
Change-Id: I9dd78cc1378051ded135a885cfc9347d42977311